Promoted editor to entire tab content area for item descriptions.
