+++ src/sheets/item/parts/ItemDescriptions.svelte
@@ -1,38 +1,18 @@
 <script lang="ts">
-  import RerenderAfterFormSubmission from 'src/components/utility/RerenderAfterFormSubmission.svelte';
-  import { FoundryAdapter } from 'src/foundry/foundry-adapter';
   import type { ItemSheetContext } from 'src/types/item';
-  import { getContext } from 'svelte';
+  import { createEventDispatcher, getContext } from 'svelte';
   import type { Readable } from 'svelte/store';
   import Accordion from 'src/components/accordion/Accordion.svelte';
   import AccordionItem from 'src/components/accordion/AccordionItem.svelte';
-  import OpenSheetEditor from 'src/components/editor/OpenSheetEditor.svelte';
 
   let context = getContext<Readable<ItemSheetContext>>('context');
 
-  const localize = FoundryAdapter.localize;
-
-  function onEditorActivation(node: HTMLElement) {
-    if (editorIsActive) {
-      editing = false;
-      editorIsActive = false;
-      return;
-    }
-
-    $context.activateFoundryJQueryListeners(node);
-    editorIsActive = true;
-  }
-
-  let editing = false;
-  let editorIsActive = false;
-  let valueToEdit: string;
-  let fieldToEdit: string;
-
-  function edit(value: string, field: string) {
-    valueToEdit = value;
-    fieldToEdit = field;
-    editing = true;
-  }
+  const dispatcher = createEventDispatcher<{
+    edit: {
+      valueToEdit: string;
+      fieldToEdit: string;
+    };
+  }>();
 
   let accordionItemOpenStates = $context.itemDescriptions.map(
     (_, i) => i === 0
@@ -40,7 +20,7 @@
 </script>
 
 <div class="item-descriptions-container">
-  <Accordion multiple class={editing ? 'hidden' : ''}>
+  <Accordion multiple>
     {#each $context.itemDescriptions as itemDescription, i (itemDescription.field)}
       <AccordionItem bind:open={accordionItemOpenStates[i]}>
         <span
@@ -54,8 +34,10 @@
               type="button"
               class="inline-icon-button edit-item-description"
               on:click|stopPropagation={() =>
-                edit(itemDescription.content, itemDescription.field)}
-              ><i class="fas fa-edit" /></button
+                dispatcher('edit', {
+                  valueToEdit: itemDescription.content,
+                  fieldToEdit: itemDescription.field,
+                })}><i class="fas fa-edit" /></button
             >
           {/if}
         </span>
@@ -63,14 +45,6 @@
       </AccordionItem>
     {/each}
   </Accordion>
-
-  {#if editing}
-    <RerenderAfterFormSubmission andOnValueChange={valueToEdit}>
-      <article class="editor-container" use:onEditorActivation>
-        <OpenSheetEditor content={valueToEdit} target={fieldToEdit} />
-      </article>
-    </RerenderAfterFormSubmission>
-  {/if}
 </div>
 
 <style lang="scss">
